Fix some bugs
- Update `func` CLI to the latest version
- Adjust to API change for fetching refs in the vscode-git extension API
